sidecar-locals-report now lists files, ignoring directory trust
Without this, reporting on a file that didn't yet trust it's
`.sidecar-locals` directory would show nothing - making it difficult
to setup sidecar-locals for new projects. Where the user intents
to set the trusted directories later.

Now show all files, ignoring trust status, leaving it up to the user
to trust the directories in the future.